+++ src/models.rs
@@ -159,7 +159,7 @@ pub struct JsonTransaction {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SubaddressBalanceData {
     pub address: Address,
-    pub address_index: u64,
+    pub address_index: u32,
     pub balance: u64,
     pub label: String,
     pub num_unspent_outputs: u64,
@@ -205,7 +205,7 @@ pub struct TransferData {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SubaddressData {
     pub address: Address,
-    pub address_index: u64,
+    pub address_index: u32,
     pub label: String,
     pub used: bool,
 }
@@ -270,8 +270,8 @@ pub struct IncomingTransfer {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SweepAllArgs {
     pub address: Address,
-    pub account_index: u64,
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub account_index: u32,
+    pub subaddr_indices: Option<Vec<u32>>,
     pub priority: TransferPriority,
     pub mixin: u64,
     pub ring_size: u64,
@@ -299,8 +299,8 @@ pub struct SweepAllData {
 /// Argument type of wallet `transfer`.
 #[derive(Clone, Debug, Default)]
 pub struct TransferOptions {
-    pub account_index: Option<u64>,
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub account_index: Option<u32>,
+    pub subaddr_indices: Option<Vec<u32>>,
     pub mixin: Option<u64>,
     pub ring_size: Option<u64>,
     pub unlock_time: Option<u64>,
@@ -323,7 +323,7 @@ pub struct GenerateFromKeysArgs {
 /// Return sub-type of wallet `get_accounts`.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct GotAccount {
-    pub account_index: u64,
+    pub account_index: u32,
     pub balance: u64,
     pub base_address: monero::Address,
     pub label: Option<String>,
@@ -394,9 +394,9 @@ impl From<GetTransfersCategory> for &'static str {
 pub struct GetTransfersSelector {
     pub category_selector: HashMap<GetTransfersCategory, bool>,
     /// Index of the account to query for transfers. (defaults to 0)
-    pub account_index: Option<u64>,
+    pub account_index: Option<u32>,
     /// List of subaddress indices to query for transfers. (Defaults to empty - all indices)
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub subaddr_indices: Option<Vec<u32>>,
     /// Filter transfers by block height.
     pub block_height_filter: Option<BlockHeightFilter>,
 }
